Create README.md
In this, we will create a CRUD API that Creates, Reads, Updates and Deletes items from a DynamoDB table. The API will run serverless, so there is no management of the underlying infrastructure and scaling is done automatically.

It features the creation of a DynamoDB table, HTTP API Gateway, Lambda function.